requirements.txt
lxml
cssselect
zope.dottedname
Pillow
requests
